On macOS, run Swank in screen/tmux, if detected
Prior to this change, Slimv starts Swank in a new xterm or Terminal
window on macOS by default even if Vim or MacVim is launched from screen
or tmux. The precise default behavior on macOS is described below:

  - If has('macunix') = 0, then Slimv runs Swank in xterm, if available.
  - If has('macunix') = 1, then Slimv runs Swank in Terminal.

This default behavior is inconsistent with the default behavior of Slimv
on Linux desktop. On Linux desktop, if we launch Vim or GVim from screen
or tmux, then Slimv starts Swank in a new screen or tmux window,
respectively, by default. Only when Vim or GVim is launched neither from
screen nor from tmux, Slimv starts Swank in a new xterm window. The same
default behavior should be followed on macOS too for the sake of
consistency.

With this change, on macOS, if we launch Vim or MacVim from screen or
tmux, then Slimv starts Swank in a new screen or tmux window,
respectively, by default. Only when Vim or MacVim is launched neither
from screen nor from tmux, Slimv starts Swank in a new Terminal window.
This default behavior is now consistent with the default behavior of
Slimv on Linux desktop.