Update CompilerException args for Clojure 1.5.0
In Clojure 1.5.0 and later versions, the constructor for
`clojure.lang.Compiler$CompilerException` requires an additional
argument that specifies column number. This change updates Swank Clojure
to check the current Clojure version and then instantiate this class
with the correct number of arguments.

This change is based on commit 12e96d0 of technomancy/swank-clojure.
Visit https://github.com/technomancy/swank-clojure/commit/12e96d0 to see
that commit.