Set classpath delimiter as per OS to run Clojure
Prior to this change, even if Clojure is available at one of the
auto-detected paths on Linux or Unix, e.g., `~/clojure/clojure.jar`, the
normal mode command `,c` fails to start Swank server with Clojure. This
issue occurs because semicolon (`;`) is used as the delimiter in the
classpath value for `java -cp` option. Here is an example classpath
value that is used prior to this change:

    /home/x/clojure/clojure.jar;/home/x/.vim/bundle/slimv/swank-clojure

Semicolon (`;`) is the right delimiter for Java running on Windows but
it should be colon (`:`) on a Linux or Unix system. This change ensures
that the correct delimiter is chosen as per the current operating
system.