Set PATH delimiter as per OS to search for Clojure
Prior to this change, on Linux or Unix, even if Clojure is available in
one of the paths mentioned in the PATH environment variable, the normal
mode command `,c` fails to start Swank server with Clojure. This issue
occurs because semicolon (`;`) is assumed to be the path delimiter in
the PATH value, so when PATH is, say, `/usr/bin:/usr/local/bin` and
`clojure.jar` is available in `/usr/local/bin`, the search for Clojure
fails because Slimv tries to search for Clojure in a single directory
named `/usr/bin:/usr/local/bin`. Instead, it should search for Clojure
in two directories named `/usr/bin` and `/usr/local/bin`.

This change ensures that on Linux or Unix, colon (`:`) is interpreted as
the path delimiter in the PATH value. On Windows, semicolon (`;`) is
used as the path delimiter.